Add PBD::ClassTracker and associated classes and macros
Add some functionality for tracking class instance counts and getting stack
traces. This is mostly for leak detection at this stage and if enabled for a
class will print the number of instances leaked at exit and stack traces where
the instances were leaked.

The functionality is enabled by compiling with --class-tracking which will
define the PBD_CLASS_TRACKER macro which when used will add a class member to a
class to track instance creation/copying and destruction.

As there is no UI for controlling the ClassTrackerManager yet the only way to
control which classes are tracked is via the environment variable
PBD_CLASS_TRACKER

For instance to track instances of class PBD::XMLNode

$ PBD_CLASS_TRACKER=XMLNode ./ardev

Extra output about when class instances are created or destroyed can be enabled
via the debug log and several classes can be specified.

$ PBD_CLASS_TRACKER="XMLNode,XMLProperty" PBD_DEBUG=classtracker ./ardev

There is a lot of potential to add and extend this sort of functionality to
become some sort of "Developer Tools" UI etc but I think for the moment this
should be useful as is.