Use consistent naming in SilentTrimmer class
The term "frame" in the context of audio data in the Ardour codebase is always
used to refer to a single audio sample. So change the name back to what it was
previously and use the term "sample" to differentiate between frames and an
interleaved "sample" that is non/silent in each channel.